backend/requirements.txt
# Core ML and Data Science Libraries
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.0.74
Pillow==10.0.0

# Web Framework
Flask==2.3.2
Flask-CORS==4.0.0

# Image Processing
imageio==2.31.1
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
joblib==1.3.2
python-dotenv==1.0.0
werkzeug==2.3.6

# Development and Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Production
gunicorn==21.0.1